Adaptação de compatibilidade
TTodas as referências restantes de "Int" na chave primária e métodos relacionados foram substituidos por "long"

+++ src/br/ufac/academico/entity/Banco.java
@@ -19,7 +19,7 @@ public class Banco {
 		
 	}
 
-	public int getCodigo() {
+	public long getCodigo() {
 		return codigo;
 	}
